+++ src/main/java/nextstep/exception/ReservationException.java
@@ -1,4 +1,4 @@
-package nextstep;
+package nextstep.exception;
 
 public class ReservationException extends RuntimeException {
